% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advRlab3package-package.R
\docType{package}
\name{advRlab3package-package}
\alias{advRlab3package}
\alias{advRlab3package-package}
\title{advRlab3package: A package for the Euclidean algorithm, and the Dijkstra algorithm}
\description{
This package contains two functions (euclidean and dijkstra) that respectively perform the Euclidean and Dijkstra algorithms.
}
\author{
\strong{Maintainer}: Simon Jorstedt \email{jorstedtsimon@gmail.com}

Authors:
\itemize{
  \item Muditha Cherangani \email{mcherangani@gmail.com}
}

}
\keyword{internal}
