% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra's algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A data.frame with variables v1, v2, w containing numeric elements.}

\item{init_node}{A numeric element.}
}
\value{
A numeric vector.
}
\description{
Dijkstra's algorithm
}
\examples{
graph <- data.frame(v1=c(...), v2=c(...), w=c(...))
dijkstra(graph, 2)
}
