% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Wiki graph}
\format{
\subsection{\code{wiki_graph}}{

A data frame with 18 rows and 3 columns:
\describe{
\item{v1}{Start node of an edge in the graph.}
\item{v2}{End node of an edge in the graph.}
\item{w}{Weight of the edge.}
}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
\usage{
wiki_graph
}
\description{
A network graph taken from \href{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}{the Dijkstra algorithms wikipedia page}.
}
\keyword{datasets}
