% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean Function}
\source{
\href{https://en.wikipedia.org/wiki/Euclidean_algorithm}{The wikipedia page of the Euclidean Algorithm}
}
\usage{
euclidean(val1, val2)
}
\arguments{
\item{val1}{A numeric value that should be compared with \code{val1}.}

\item{val2}{A numeric value that should be compared with \code{val2}.}
}
\value{
A numeric: the greatest common divisor of \code{val1} and \code{val2}.
}
\description{
The Euclidean algorithm is a way to find the Greatest Common Divisor (GCD) of two positive integers.
}
