% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean Function}
\usage{
euclidean(val1, val2)
}
\arguments{
\item{val1}{A numeric value that should be compared with \code{val1}.}

\item{val2}{A numeric value that should be compared with \code{val2}.}
}
\value{
A numeric: the greatest common divisor of \code{val1} and \code{val2}.
}
\description{
In mathematics, the Euclidean algorithm or Euclid's algorithm, is an efficient method for computing the greatest common divisor (GCD) of two integers (numbers), the largest number that divides them both without a remainder.
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)
}
